+++ component-library/src/rich-text-editor/components/index.tsx
