--- component-library/src/rich-text-editor/plugins.tsx
@@ -1,34 +1,0 @@
-import {
-  createBlockquotePlugin,
-  createBoldPlugin,
-  createCodeBlockPlugin,
-  createCodePlugin,
-  createHeadingPlugin,
-  createHistoryPlugin,
-  createItalicPlugin,
-  createListPlugin,
-  createParagraphPlugin,
-  createReactPlugin,
-  createStrikethroughPlugin,
-  createUnderlinePlugin
-} from '@udecode/slate-plugins'
-
-export const pluginsBasic = [
-  // editor
-  createReactPlugin(),
-  createHistoryPlugin(),
-
-  // elements
-  createParagraphPlugin(),
-  createBlockquotePlugin(),
-  createCodeBlockPlugin(),
-  createHeadingPlugin(),
-  createListPlugin(),
-
-  // marks
-  createBoldPlugin(),
-  createItalicPlugin(),
-  createUnderlinePlugin(),
-  createStrikethroughPlugin(),
-  createCodePlugin()
-]
